add updated changes

+++ src/main/java/com/pearl/warehouse/dto/input/ProductInput.java
@@ -1,4 +1,4 @@
 package com.pearl.warehouse.dto.input;
 
-public record ProductInput(Integer id,String productName,String code,Integer categoryId) {
+public record ProductInput(Long id,String productName,String code,Integer categoryId) {
 }
